+++ packages/react-native-reanimated/src/common/style/registry.ts
@@ -9,10 +9,8 @@ import {
 } from './propsBuilder';
 
 const DEFAULT_SEPARATELY_INTERPOLATED_NESTED_PROPERTIES = new Set<string>([
-  'boxShadow',
   'shadowOffset',
   'textShadowOffset',
-  'transformOrigin',
 ]);
 
 type PropsBuilderEntry = {